Update compute_forest to account for 3 or 4 bands

+++ forestatrisk/data/country_compute.py
@@ -8,9 +8,9 @@ import numpy as np
 from osgeo import gdal
 
 from ..misc import make_dir
-from .compute import compute_biomass_avitabile, compute_osm
-from .compute import compute_srtm, compute_wdpa
 from .compute import compute_forest
+from .compute import compute_srtm, compute_wdpa, compute_osm
+from .compute import compute_biomass_avitabile
 from .get_vector_extent import get_vector_extent
 
 
@@ -103,30 +103,33 @@ def country_compute(
         wd = os.getcwd()
         os.chdir(temp_dir)
         # Perform computations
-        compute_forest(iso3, proj, extent_reg)
+        compute_forest(proj, extent_reg)
         # Create directories
         make_dir(os.path.join(wd, output_dir, "forest"))
         make_dir(os.path.join(wd, output_dir, "validation"))
         make_dir(os.path.join(wd, output_dir, "forecast"))
         # Copy files for modelling
-        ifiles = ["dist_edge_t1.tif", "fcc12.tif"]
-        ofiles = ["dist_edge.tif", "fcc.tif"]
+        ifiles = ["dist_edge_t1.tif", "dist_defor_t1.tif", "fcc12.tif"]
+        ofiles = ["dist_edge.tif", "dist_defor.tif", "fcc.tif"]
         for (ifile, ofile) in zip(ifiles, ofiles):
-            copy2(ifile, os.path.join(wd, output_dir, ofile))
+            if os.path.isfile(ifile):
+                copy2(ifile, os.path.join(wd, output_dir, ofile))
         # Copy files for validation
         files = ["dist_edge_t2.tif", "dist_defor_t2.tif"]
         for file in files:
-            copy2(file, os.path.join(wd, output_dir, "validation"))
+            if os.path.isfile(file):
+                copy2(file, os.path.join(wd, output_dir, "validation"))
         # Copy files for forecast
         files = ["dist_edge_t3.tif", "dist_defor_t3.tif"]
         for file in files:
-            copy2(file, os.path.join(wd, output_dir, "forecast"))
+            if os.path.isfile(file):
+                copy2(file, os.path.join(wd, output_dir, "forecast"))
         # Forest data
-        files = ["forest_t1.tif", "forest_t2.tif", "forest_t3.tif",
-                 "forest_2005.tif", "forest_2015.tif", "fcc23.tif",
-                 "fcc123.tif", "fcc12345.tif"]
+        files = [f"forest_t{i}.tif" for i in range(4)]
+        files.append("fcc123.tif")
         for file in files:
-            copy2(file, os.path.join(wd, output_dir, "forest"))
+            if os.path.isfile(file):
+                copy2(file, os.path.join(wd, output_dir, "forest"))
         os.chdir(wd)
 
     # Keep or remove directory
